feat: Add data parser

+++ program/src/dtw.h
@@ -3,6 +3,6 @@
 
 #include <vector>
 
-double calculateDTWDistance(std::vector<std::vector<int> > &a, std::vector<std::vector<int> > &b);
+double calculateDTWDistance(std::vector<std::vector<double> > &a, std::vector<std::vector<double> > &b);
 
 #endif
